fix(settings): keep the close-deadline migration alive on hand-edited values
- 值改以 numeric 夾擠後再轉 int:1.5 或超出 int4 範圍的髒值原本會讓整支遷移回滾
- 期限推導三個消費點一律 int(lock_days)
- 遷移測試釘明確 revision,不隨新 revision 落地而失準

+++ backend/alembic/versions/a3f7c2e91b48_close_lock_in_days.py
@@ -30,7 +30,7 @@ def upgrade() -> None:
             """
             UPDATE system_settings
             SET key = 'close_lock_days',
-                value = to_jsonb(GREATEST(1, LEAST(366, (value #>> '{}')::int * :per_month)))
+                value = to_jsonb(GREATEST(1, LEAST(366, (value #>> '{}')::numeric * :per_month))::int)
             WHERE key = 'close_lock_months' AND jsonb_typeof(value) = 'number'
             """
         ),
@@ -51,7 +51,7 @@ def downgrade() -> None:
             UPDATE system_settings
             SET key = 'close_lock_months',
                 value = to_jsonb(
-                    LEAST(6, GREATEST(1, CEIL((value #>> '{}')::numeric / :per_month)::int))
+                    LEAST(6, GREATEST(1, CEIL((value #>> '{}')::numeric / :per_month))::int)
                 )
             WHERE key = 'close_lock_days' AND jsonb_typeof(value) = 'number'
             """
